added a map loader
